complete re-do
