+++ src/state/auto-scroller/fluid-scroller/get-required-scroll/get-on-axis/get-pixel-thresholds.js
@@ -1,0 +1,22 @@
+// @flow
+import type { Rect } from 'css-box-model';
+import config from '../../config';
+import type { Axis } from '../../../../../types';
+
+export type PixelThresholds = {|
+  startFrom: number,
+  maxSpeedAt: number,
+|};
+
+// converts the percentages in the config into actual pixel values
+export default (container: Rect, axis: Axis): PixelThresholds => {
+  const startFrom: number = container[axis.size] * config.startFrom;
+  const maxSpeedAt: number = container[axis.size] * config.maxSpeedAt;
+
+  const thresholds: PixelThresholds = {
+    startFrom,
+    maxSpeedAt,
+  };
+
+  return thresholds;
+};
